fix: move all period handling to OrgUnitData
